swls_core

Module components

Defines all common Components and Resources

In this World, Entitys are documents and Components are derived from these documents. Different Systems derive new Components from existing Components, that are added to the Entity. For example, if Triples are defined, [systems::derive_classes] will derive DefinedClass from them and add them to the Entity.

Re-exports

pub use config::*;

pub use document::*;

pub use resources::*;

pub use type_hierarchy::*;

Modules

config

document

resources

type_hierarchy