swls_core

Module prelude

Re-exports

pub use completion::Label as CompletionLabel;

pub use hover::Label as HoverLabel;

pub use parse::Label as ParseLabel;

pub use rename::PrepareRename as PrepareRenameLabel;

pub use rename::Rename as RenameLabel;

pub use diagnostics::Label as DiagnosticsLabel;

pub use save::Label as SaveLabel;

pub use inlay::Label as InlayLabel;

pub use format::Label as FormatLabel;

pub use semantic::Label as SemanticLabel;

pub use references::Label as ReferencesLabel;

pub use goto_definition::Label as GotoDefinitionLabel;

pub use goto_type::Label as GotoTypeLabel;

pub use code_action::Label as CodeActionLabel;

pub use config::*;

pub use document::*;

pub use resources::*;

pub use type_hierarchy::*;

Modules

code_action

completion

config

diagnostics

document

feature

Hosts all common features of the semantic language server. All features supported by the language server. Features like completion (CompletionRequest) and hover expose a component that collects items returned to the client.

format

goto_definition

goto_type

hover

inlay

parse

references

rename

resources

save

semantic

semantic_token

systems

type_hierarchy

Structs

Backend

CompletionRequest

Component indicating that the current document is currently handling a Completion request.

CreateEvent

Event triggers when a document is opened

DiagnosticItem

DiagnosticPublisher

DiagnosticSender

File

FormatRequest

Component indicating that the current document is currently handling a Format request.

Fs

HighlightRequest

Component indicating that the current document is currently handling a Hightlight request.

HoverRequest

Component indicating that the current document is currently handling a Hover request.

InlayRequest

Component indicating that the current document is currently handling a Inlay request.

MyQuad

MyTerm

PToken

Token struct holding the token and the index in the token array

Prefix

One defined prefix, maps prefix to url

Prefixes

Component that containing defined prefixes and base URL.

PrepareRenameRequest

Component indicating that the current document is currently handling a PrepareRename request.

RenameEdits

Component indicating that the current document is currently handling a Rename request, collecting TextEdits.

SemanticTokensDict

Resource mapping a [‘SemanticTokenType’] to their used index.

SimpleCompletion

SimpleDiagnostic

Spanned

TokenComponent

Component used to indicate the currently targeted Token during a request.

Tokens

Component that contains the parsed tokens.

TripleComponent

Component used to indicate the currently targeted MyQuad<'static> during a request.

Triples

Component containing all derived Triples from the documents.

Triples2

Enums

SparqlAggregate

SparqlCall

SparqlExpr

SparqlExpr2

SparqlExpr3

SparqlKeyword

StringStyle

TermContext

Token

TripleTarget

Component used to indicate the term type of currently targeted Token in the Triple.

Traits

Client

ClientSync

FsTrait

Lang

LangHelper

Membered

TokenTrait

Functions

get_current_token

get_current_triple

lsp_range_to_range

offset_to_position

offsets_to_range

position_to_offset

range_to_range

setup_schedule_labels

Initializes a World, including Resources and Schedules. All systems defined in crate are added to the World.

spanned

spawn_or_insert