swls_core::prelude

Module document

Structs

CurrentType

Dirty

Component indicating whether or not the document is dirty, a dirty document parsed with errors.

DocumentLinks

Component used to remember the linked documents.

DynLang

Component that allows for language specific implementation for certain things, reducing code duplication.

Element

Component that contains the parsed semantic element (i.e. Turtle, JSONLD).

Errors

Simple wrapper for errors that derives Component

Global

Indicates that this should be a global entity, linked to all other documents

KeyWords

Component containing the typical keywords for the current language.

Label

Component containing the [lsp_types::Url] of the current document.

Open

Component indicating whether or not the document is actually open.

PositionComponent

Component used to wrap an incoming [lsp_types::Position].

RopeC

Component containing the current source code as ropey::Rope

Source

Component containing the current source code as String

Wrapped

Simple wrapper structure that derives Component