swls_core

Module systems

Modules

prefix

Structs

DefinedClass

DefinedProperty

FromPrefix

OntologyExtractor

PrefixEntry

TypeId

Functions

check_added_ontology_extract

complete_class

complete_properties

derive_ontologies

derive_owl_imports_links

derive_prefix_links

derive_shapes

extract_type_hierarchy

fetch_lov_properties

First of all, fetch the lov dataset information at url https://lov.linkeddata.es/dataset/lov/api/v2/vocabulary/info?vocab=${prefix} Next, extract that json object into an object and find the latest dataset

handle_tasks

hover_class

hover_property

hover_types

infer_current_type

infer_types

init_ontology_extractor

keyword_complete

open_imports

populate_known_ontologies

spawn_or_insert

validate_shapes

System evaluates linked shapes

Type Aliases

DefinedClasses

DefinedProperties